SIMServer/src/main/java/sim/utils/JsonUtils.java
package sim.utils;

import java.util.List;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.JavaType;
import com.fasterxml.jackson.databind.ObjectMapper;

/**
 * Custom response structure, conversion class
 */
public class JsonUtils {

    // 定义jackson对象
    private static final ObjectMapper MAPPER = new ObjectMapper();

    /**
     * transform object into json
     * <p>Title: pojoToJson</p>
     * <p>Description: </p>
     *
     * @param data the data
     * @return String string
     */
    public static String objectToJson(Object data) {
    	try {
			String string = MAPPER.writeValueAsString(data);
			return string;
		} catch (JsonProcessingException e) {
			e.printStackTrace();
		}
    	return null;
    }

    /**
     * transform json to object
     *
     * @param <T>      the type parameter
     * @param jsonData json数据
     * @param beanType the bean type
     * @return t
     */
    public static <T> T jsonToPojo(String jsonData, Class<T> beanType) {
        try {
            T t = MAPPER.readValue(jsonData, beanType);
            return t;
        } catch (Exception e) {
        	e.printStackTrace();
        }
        return null;
    }

    /**
     * list transform json into pojo object
     * <p>Title: jsonToList</p>
     * <p>Description: </p>
     *
     * @param <T>      the type parameter
     * @param jsonData the json data
     * @param beanType the bean type
     * @return list
     */
    public static <T>List<T> jsonToList(String jsonData, Class<T> beanType) {
    	JavaType javaType = MAPPER.getTypeFactory().constructParametricType(List.class, beanType);
    	try {
    		List<T> list = MAPPER.readValue(jsonData, javaType);
    		return list;
		} catch (Exception e) {
			e.printStackTrace();
		}
    	
    	return null;
    }
    
}
